Almost there...

+++ modelio.py
@@ -71,13 +71,13 @@ def prepare_data(seqs, labels, maxlen=None):
     n_samples = len(seqs)
     maxlen = numpy.max(lengths)
 
-    x = numpy.zeros((maxlen, n_samples)).astype('int64')
+    x = numpy.zeros((maxlen, n_samples)).astype('int8')
     x_mask = numpy.zeros((maxlen, n_samples)).astype(theano.config.floatX)
     #words_mask = numpy.zeros((maxlen, 80, n_samples)).astype(theano.config.floatX)
 #   words_mask = numpy.zeros((maxlen * n_samples, 4)).astype('int64')
     words_mask = []
-    y = numpy.zeros((16, n_samples)).astype('int64')
-    y_mask = numpy.zeros((16, n_samples)).astype('int32')
+    y = numpy.zeros((16, n_samples)).astype('int8')
+    y_mask = numpy.zeros((16, n_samples)).astype('int8')
     for idx, (s, l) in enumerate(zip(seqs, labels)):
         # idx is the current posdiction in the mini-batch
         # s is a list of characters
@@ -110,7 +110,7 @@ def prepare_data(seqs, labels, maxlen=None):
             y_mask[c, idx] = 1
             i += 1
 
-    words_mask = numpy.asarray(words_mask, dtype='int32')
+    words_mask = numpy.asarray(words_mask, dtype='int8')
 #    print
 
     return x, x_mask, words_mask, y, y_mask
